+++ server/s3_option.h
@@ -117,7 +117,7 @@ class S3Option {
   int audit_logger_port;
   std::string audit_logger_rsyslog_msgid;
   std::string audit_logger_kafka_web_path;
-  long audit_logger_iem_port;
+  int audit_logger_iem_port;
   std::string audit_logger_iem_path;
   unsigned short max_audit_retry_count;
   std::string s3server_ssl_cert_file;
@@ -346,7 +346,7 @@ class S3Option {
   int get_audit_logger_port();
   std::string get_audit_logger_rsyslog_msgid();
   std::string get_audit_logger_kafka_web_path();
-  long get_audit_logger_iem_port();
+  int get_audit_logger_iem_port();
   std::string get_audit_logger_iem_path();
   unsigned short get_audit_max_retry_count();
   unsigned short get_s3_bind_port();
